+++ Backend/src/users/dto/loginUser.dto.ts
@@ -1,0 +1,10 @@
+import { IsEmail, IsNotEmpty } from 'class-validator';
+
+export class LoginUserDto {
+  @IsEmail()
+  @IsNotEmpty()
+  email: string;
+
+  @IsNotEmpty()
+  password: string;
+}
