+++ Backend/src/users/types/reponse.types.ts
@@ -1,0 +1,10 @@
+export interface LoginResponse {
+  statusCode: number;
+  message: string;
+  data: { accessToken: string };
+}
+
+export interface ErrorResponse {
+  statusCode: number;
+  error: string;
+}
